Extract `submitOnEnter`, replace placeholder with label
